revert back to using Gzip; synchronize on write to prevent the cache trying to read buckets that are being written

+++ src/main/java/com/github/jillesvangurp/persistentcachingmap/PersistentCachingMap.java
@@ -38,9 +38,10 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.zip.GZIPInputStream;
+import java.util.zip.GZIPOutputStream;
 
-import org.apache.commons.compress.compressors.bzip2.BZip2CompressorInputStream;
-import org.apache.commons.compress.compressors.bzip2.BZip2CompressorOutputStream;
 import org.apache.commons.lang.StringUtils;
 
 import com.google.common.cache.CacheBuilder;
@@ -114,7 +115,8 @@ public class PersistentCachingMap<Key,Value> implements Iterable<Map.Entry<Key,V
                 return cache.get(bucketId).get((Key)key);
             }
         } catch (ExecutionException e) {
-            throw new IllegalStateException(e);
+            e.printStackTrace();
+            return null;
         }
         return null;
     }
@@ -174,7 +176,7 @@ public class PersistentCachingMap<Key,Value> implements Iterable<Map.Entry<Key,V
                 }
             }
 
-            try(BufferedWriter out = bzipFileWriter(path)) {
+            try(BufferedWriter out = gzipFileWriter(path)) {
                 for(Long id: bucketIds) {
                     out.write("" + id +"\n");
                 }
@@ -187,7 +189,7 @@ public class PersistentCachingMap<Key,Value> implements Iterable<Map.Entry<Key,V
     void readBucketIds() {
         File bucketIdFile = new File(dataDir,"bucketIds.gz");
         if(bucketIdFile.exists()) {
-            try(LineIterable iterable = new LineIterable(bzipFileReader(bucketIdFile))) {
+            try(LineIterable iterable = new LineIterable(gzipFileReader(bucketIdFile))) {
                 for(String line: iterable) {
                     bucketIds.add(Long.valueOf(line.trim()));
                 }
@@ -276,16 +278,24 @@ public class PersistentCachingMap<Key,Value> implements Iterable<Map.Entry<Key,V
     }
 
 
-    private static BufferedWriter bzipFileWriter(File file) throws IOException {
-        return new BufferedWriter(new OutputStreamWriter(new BZip2CompressorOutputStream(new FileOutputStream(file)), Charset.forName("utf-8")),64*1024);
+//    private static BufferedWriter bzipFileWriter(File file) throws IOException {
+//        return new BufferedWriter(new OutputStreamWriter(new BZip2CompressorOutputStream(new FileOutputStream(file)), Charset.forName("utf-8")),64*1024);
+//    }
+//
+//    private static BufferedReader bzipFileReader(File file) throws IOException {
+//        return new BufferedReader(new InputStreamReader(new BZip2CompressorInputStream(new FileInputStream(file)),Charset.forName("utf-8")));
+//    }
+    private static BufferedWriter gzipFileWriter(File file) throws IOException {
+        return new BufferedWriter(new OutputStreamWriter(new GZIPOutputStream(new FileOutputStream(file)), Charset.forName("utf-8")),64*1024);
     }
 
-    private static BufferedReader bzipFileReader(File file) throws IOException {
-        return new BufferedReader(new InputStreamReader(new BZip2CompressorInputStream(new FileInputStream(file)),Charset.forName("utf-8")));
+    private static BufferedReader gzipFileReader(File file) throws IOException {
+        return new BufferedReader(new InputStreamReader(new GZIPInputStream(new FileInputStream(file)),Charset.forName("utf-8")));
     }
 
     public class Bucket implements Iterable<Map.Entry<Key, Value>> {
         long id;
+        AtomicBoolean changed=new AtomicBoolean(false);
 
         Map<Key, Value> map = new ConcurrentHashMap<>();
 
@@ -303,33 +313,41 @@ public class PersistentCachingMap<Key,Value> implements Iterable<Map.Entry<Key,V
 
         public void put(Key key, Value value) {
             map.put(key, value);
+            changed.set(true);
         }
 
         public Value remove(Key key) {
+            changed.set(true);
             return map.remove(key);
         }
 
         public void write() {
-            File path = bucketPath(id);
-            if (map.isEmpty() && path.exists()) {
-                // map is empty, remove the file if it existed and delete the bucket from the bucketIds
-                if (!path.delete()) {
-                    throw new IllegalStateException("could not delete " + path);
-                }
-                bucketIds.remove(id);
-            } else {
-                File dir = path.getParentFile();
-                if (!dir.exists()) {
-                    if (!dir.mkdirs()) {
-                        throw new IllegalStateException("could not create directory " + dir);
+            if(changed.get()) {
+                synchronized(cache) {
+                    // prevent somebody reading a partially written bucket by synchronizing on the cache
+                    // need more clever synchronization
+                    File path = bucketPath(id);
+                    if (map.isEmpty() && path.exists()) {
+                        // map is empty, remove the file if it existed and delete the bucket from the bucketIds
+                        if (!path.delete()) {
+                            throw new IllegalStateException("could not delete " + path);
+                        }
+                        bucketIds.remove(id);
+                    } else {
+                        File dir = path.getParentFile();
+                        if (!dir.exists()) {
+                            if (!dir.mkdirs()) {
+                                throw new IllegalStateException("could not create directory " + dir);
+                            }
+                        }
+                        try (BufferedWriter out = gzipFileWriter(path)) {
+                            for (Entry<Key, Value> e : map.entrySet()) {
+                                out.write("" + codec.serializeKey(e.getKey()) + ";" + codec.serializeValue(e.getValue()) + "\n");
+                            }
+                        } catch (IOException e) {
+                            throw new IllegalStateException("could not write file " + path, e);
+                        }
                     }
-                }
-                try (BufferedWriter out = bzipFileWriter(path)) {
-                    for (Entry<Key, Value> e : map.entrySet()) {
-                        out.write("" + codec.serializeKey(e.getKey()) + ";" + codec.serializeValue(e.getValue()) + "\n");
-                    }
-                } catch (IOException e) {
-                    throw new IllegalStateException("could not write file " + path, e);
                 }
             }
         }
@@ -337,7 +355,7 @@ public class PersistentCachingMap<Key,Value> implements Iterable<Map.Entry<Key,V
         public void read() {
             File path = bucketPath(id);
             if(path.exists()) {
-                try(LineIterable iterable = new LineIterable(bzipFileReader(path))) {
+                try(LineIterable iterable = new LineIterable(gzipFileReader(path))) {
                     for(String line: iterable) {
                         int idx = line.indexOf(';');
                         Key key = codec.deserializeKey(line.substring(0,idx));
